added styling to navbar, productcard, and product page. Connected product page with product card
