+++ src/components/ProductCard.js
@@ -4,28 +4,29 @@ import React from 'react';
 import img1 from '../images/chocolate_cake.jpg'; 
 import "bootstrap/dist/css/bootstrap.min.css";
 import {Container, Row, Col} from "react-bootstrap";
+import {Link} from "react-router-dom";
 import StarRating from './StarRating'
+import "./ProductCard.css";
 
 const ProductCard = (props) => {
-
-
+    const productLink = `product/${props.id}`;
     return( 
-        <Card style = {{width: '18rem'}}>
-            <Card.Img variant = "top" src = {img1}/>
-            <Card.Body>
-                <Container className = "mt-100">
+        <Card className = 'cardWidth cardShadowHover'>
+            <Card.Img variant = "top" src = {img1} className = 'imgHeight'/>
+            <Card.Body className = "p-0">
+                <Container className = "p-0">
                     <Card.Title>{props.title}</Card.Title>
                     <Row>
-                        <Col>
+                        <Col className = "me-1">
                             <Card.Text>{props.price}</Card.Text>
                         </Col>
                         <Col>
-                            <StarRating/>
+                            <StarRating rating = "5"/>
                         </Col>
                     </Row>
                     <Card.Text>{props.shop}</Card.Text>
                 </Container>
-                <a href = "/product/${props.id}" className = "stretched-link" data-testid = "link"></a>
+                <Link to = {productLink} className = "stretched-link" data-testid = "link"/>
             </Card.Body>
         </Card>
     );
